backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
coverage==7.7.1
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
maturin==1.8.3
motor==3.3.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.13.0
uvicorn==0.27.1
